Inverts the fenced code block delimiters
Previously, the section explaining block level annotation for fenced code blocks used tildes as the code block delimiter.
This was because backticks were used in the outer code block that contained the example. Tildes are not the most common
way to specify code blocks so this commit swaps the inner and outer delimiters so that the example uses the more common
backtick syntax.